make the sign up form reusable
